requirements.txt
pandas>=2.1
plotly>=5.17
streamlit>=1.31
psycopg2-binary>=2.9
streamlit-authenticator
altair==4.2.2
python-dotenv

